+++ packages/core/src/components/record/Record.tsx
@@ -1,7 +1,7 @@
 import { BrandingOverlayType, Field } from '@bifold/oca/build/legacy'
 import React, { useCallback, useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { FlatList, StyleSheet, TouchableOpacity, View } from 'react-native'
+import { FlatList, StyleSheet, TouchableOpacity, View, Text } from 'react-native'
 
 import { useTheme } from '../../contexts/theme'
 import { testIdWithKey } from '../../utils/testable'
@@ -18,21 +18,23 @@ export interface RecordProps {
   fields: Field[]
   hideFieldValues?: boolean
   field?: (field: Field, index: number, fields: Field[]) => React.ReactElement | null
+  scrollEnabled?: boolean
 }
 
-const Record: React.FC<RecordProps> = ({ header, footer, fields, hideFieldValues = false, field = null }) => {
+const Record: React.FC<RecordProps> = ({ header, footer, fields, hideFieldValues = false, field = null, scrollEnabled = true }) => {
   const { t } = useTranslation()
   const [shown, setShown] = useState<boolean[]>([])
-  const { ListItems, TextTheme } = useTheme()
+  const { ListItems, TextTheme, Spacing } = useTheme()
   const [bundleResolver] = useServices([TOKENS.UTIL_OCA_RESOLVER])
 
   const styles = StyleSheet.create({
     linkContainer: {
       ...ListItems.recordContainer,
       flexDirection: 'row',
-      justifyContent: 'flex-end',
+      justifyContent: 'space-between',
       paddingHorizontal: bundleResolver.getBrandingOverlayType() === BrandingOverlayType.Branding10 ? 25 : 16,
-      paddingVertical: 16,
+      paddingVertical: Spacing.sm,
+      paddingBottom: Spacing.lg,
     },
     link: {
       minHeight: TextTheme.normal.fontSize,
@@ -51,6 +53,7 @@ const Record: React.FC<RecordProps> = ({ header, footer, fields, hideFieldValues
   return (
     <FlatList
       data={fields}
+      scrollEnabled={scrollEnabled}
       keyExtractor={({ name }, index) => name || index.toString()}
       showsVerticalScrollIndicator={false}
       renderItem={({ item: attr, index }) =>
@@ -76,6 +79,7 @@ const Record: React.FC<RecordProps> = ({ header, footer, fields, hideFieldValues
             {header()}
             {hideFieldValues ? (
               <View style={styles.linkContainer}>
+                <Text style={TextTheme.headingFour}>{t('ProofRequest.DetailsHeader')}</Text>
                 <TouchableOpacity
                   style={styles.link}
                   activeOpacity={1}