+++ packages/core/src/hooks/useBifoldAgentSetup.ts
@@ -119,6 +119,7 @@ const useBifoldAgentSetup = (): AgentSetupReturnType => {
         if (restartedAgent) {
           logger.info('Successfully restarted existing agent...')
           agentInstanceRef.current = restartedAgent
+          bridge.setAgent(restartedAgent)
           setAgent(restartedAgent)
           return
         }