+++ packages/react-hooks/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 3.0.3
+
+## 3.0.2
+
 ## 3.0.1
 
 ### Patch Changes