+++ packages/react-native-attestation/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @bifold/react-native-attestation
 
+## 3.0.3
+
+## 3.0.2
+
 ## 3.0.1
 
 ## 3.0.0