+++ packages/verifier/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @bifold/verifier
 
+## 3.0.3
+
+### Patch Changes
+
+- @bifold/react-hooks@3.0.3
+
+## 3.0.2
+
+### Patch Changes
+
+- @bifold/react-hooks@3.0.2
+
 ## 3.0.1
 
 ### Patch Changes